+++ src/sema/aggregate_layout.mli
@@ -40,6 +40,7 @@ type 'query generic_expression = 'query Closed_numeric_expression.expression =
   | Unsigned_integer_expression of { value : int64; origin : Symbol.origin }
   | Floating_expression of { value : float; origin : Symbol.origin }
   | Current_position_expression of Symbol.origin
+  | Captured_position_expression of Symbol.origin * int64
   | Unary_expression of {
       operator : unary_operator;
       operand : 'query generic_expression;