+++ src/sema/closed_numeric_expression.mli
@@ -36,6 +36,7 @@ type 'query expression =
   | Unsigned_integer_expression of { value : int64; origin : Symbol.origin }
   | Floating_expression of { value : float; origin : Symbol.origin }
   | Current_position_expression of Symbol.origin
+  | Captured_position_expression of Symbol.origin * int64
   | Unary_expression of {
       operator : unary_operator;
       operand : 'query expression;
@@ -111,6 +112,7 @@ val binary : Frontend.Operator.binary_operator -> binary_operator option
 
 val of_ast :
   ?allow_floating:bool ->
+  ?position_value:(Frontend.Ast.expression -> int64 option) ->
   query_expression:('q -> Frontend.Ast.expression) ->
   queries:'q list ->
   Frontend.Ast.expression ->